034_fm

+++ src/clover_sketches_lib/lib/peq/filter_response.cpp
@@ -22,19 +22,19 @@ void log_spaced_freqs(std::vector<float>& out, float start_freq, float stop_freq
 }
 
 void compute_complex_response(
-        const clover::dsp::iir_coeffs& coeffs, complex_response& out, const std::vector<float>& freqs) {
+    const clover::dsp::iir_coeffs& coeffs, complex_response& out, const std::vector<float>& freqs) {
     if (out.size() != freqs.size())
         throw std::invalid_argument(
-                std::format(
-                        //
-                        "compute_response(coeffs, out, freqs): \n"
-                        "out.size() != freqs.size()\n"
-                        "{} {}",
-                        out.size(),
-                        freqs.size()));
+            std::format(
+                //
+                "compute_response(coeffs, out, freqs): \n"
+                "out.size() != freqs.size()\n"
+                "{} {}",
+                out.size(),
+                freqs.size()));
 
     const float stop_freq = *(freqs.end() - 1);
-    for (auto [freq, output, i] : std::views::zip(freqs, out, std::views::iota(0, (int)out.size()))) {
+    for (auto [freq, output, i] : std::views::zip(freqs, out, std::views::iota(0u, out.size()))) {
         const float omega           = 2.0f * std::numbers::pi_v<float> * freq / (2.0f * stop_freq);
         const std::complex<float> z = std::polar(1.0f, -omega);
 
@@ -48,11 +48,11 @@ void compute_complex_response(
 void compute_spectrum(const complex_response& spectrum, computed_response& out) {
     if (spectrum.size() != out.magnitudes.size() || spectrum.size() != out.angles.size())
         throw std::invalid_argument(
-                std::format(
-                        "compute_spectrum(...) size mismatch {} {} {}",
-                        spectrum.size(),
-                        out.magnitudes.size(),
-                        out.angles.size()));
+            std::format(
+                "compute_spectrum(...) size mismatch {} {} {}",
+                spectrum.size(),
+                out.magnitudes.size(),
+                out.angles.size()));
 
     for (auto [z, magnitude, angle] : std::views::zip(spectrum, out.magnitudes, out.angles)) {
         magnitude = std::abs(z);
